GPU: source portability fixes for the Windows CUDA build
- buffer_comparator_kernel_lib.cu.h: use fabsf/fmaxf on float operands;
  with MSVC's math headers the double overloads are ambiguous in device
  code here.
- backends/gpu/autotuner/cudnn.cc: AsDnn() returns null when no DNN
  plugin (cuDNN) is linked into the binary, as in the first-light winjax
  plugin build; return FailedPrecondition instead of dereferencing it in
  GetConvolutionCustomCallConfigs.

+++ xla/stream_executor/gpu/buffer_comparator_kernel_lib.cu.h
@@ -109,7 +109,7 @@ __global__ void xla_int_comparison(T* buffer_a, T* buffer_b,
       elem_b = static_cast<uint64_t>(buffer_b[idx]);
     }
     float rel_error =
-        fabs(elem_a - elem_b) / (fmax(fabs(elem_a), fabs(elem_b)) + 1);
+        fabsf(elem_a - elem_b) / (fmaxf(fabsf(elem_a), fabsf(elem_b)) + 1);
     if (rel_error > rel_error_threshold || isnan(rel_error)) {
       atomicAdd(mismatch_count, 1);
     }
